requirements.txt
dataclasses-json
fastapi
requests
google-generativeai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
PyPDF2
httpcore
httptools
httpx
langchain
langchain_google_genai
faiss-cpu
langchain-text-splitters
pydantic
python-dotenv
uvicorn